pyproject.toml
[tool.poetry]
name = "pubmed-pharma-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with pharmaceutical/biotech company affiliations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pubmed_pharma_fetcher"}]

[tool.poetry.dependencies]
python = "^3.8.1"
biopython = "^1.81"
pandas = "^2.0.0"
click = "^8.1.0"
requests = "^2.31.0"
lxml = "^4.9.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
mypy = "^1.5.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
get-papers-list = "pubmed_pharma_fetcher.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
